[guides] Add start of a guide on writing API documentation
Why
---
We are beginning to autogenerate parts of our API documentation from docblocks in our code. The contents of docblocks will become more visible and will reflect the quality of our work more.

As a baseline, we want our documentation to be objectively correct. This means correct grammar and spelling and using programming terms correctly.

We also want our documentation to be subjectively useful. This means documenting APIs beyond shallow function signatures—explain the behavior of the iceberg below the surface.

There are more things we could add to this guide over time.